Add spf13/cobra lib to support CLI flags
Move the web server functionality under the "server" CLI command
