+++ .gitignore
@@ -1,2 +1,2 @@
-telefonistka
+/telefonistka
 vendor/
